Moved protobuf definitions and API to runtime
