+++ infrastructure/src/main/java/org/corfudb/infrastructure/LayoutServer.java
@@ -10,16 +10,16 @@ import javax.annotation.Nonnull;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.corfudb.common.protocol.API;
-import org.corfudb.common.protocol.proto.CorfuProtocol;
-import org.corfudb.common.protocol.proto.CorfuProtocol.BootstrapLayoutResponse;
-import org.corfudb.common.protocol.proto.CorfuProtocol.CommitLayoutResponse;
-import org.corfudb.common.protocol.proto.CorfuProtocol.Header;
-import org.corfudb.common.protocol.proto.CorfuProtocol.PrepareLayoutResponse;
-import org.corfudb.common.protocol.proto.CorfuProtocol.ProposeLayoutResponse;
-import org.corfudb.common.protocol.proto.CorfuProtocol.Request;
-import org.corfudb.common.protocol.proto.CorfuProtocol.Response;
-import org.corfudb.common.protocol.proto.CorfuProtocol.ServerError;
+import org.corfudb.protocols.API;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.BootstrapLayoutResponse;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.CommitLayoutResponse;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.Header;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.PrepareLayoutResponse;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.ProposeLayoutResponse;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.Request;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.Response;
+import org.corfudb.runtime.protocol.proto.CorfuProtocol.ServerError;
 import org.corfudb.infrastructure.ServerThreadFactory.ExceptionHandler;
 import org.corfudb.infrastructure.paxos.PaxosDataStore;
 import org.corfudb.protocols.wireprotocol.CorfuMsg;
