+++ runtime/proto/CorfuProtocol.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package org.corfudb.common;
+package org.corfudb.runtime;
 
-option java_package = "org.corfudb.common.protocol.proto";
+option java_package = "org.corfudb.runtime.protocol.proto";
 
 /**
  * Common Definitions
