+++ runtime/src/main/java/org/corfudb/protocols/wireprotocol/NettyCorfuMessageDecoder.java
@@ -6,7 +6,7 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.ByteToMessageDecoder;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.corfudb.common.protocol.API;
+import org.corfudb.protocols.API;
 
 /**
  * Created by mwei on 10/1/15.
